# internal/swagger/api_v1.0.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "GTB Swagger",
      "description": "GTB server API",
      "contact": {
        "email": "developer@energie-ip.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "/{basePath}",
        "variables": {
          "basePath": {
            "default": "v1.0"
          }
        }
      }
    ],
    "paths": {
...
      "/setup/hvac/{mac}": {
        "get": {
          "tags": [
            "setup"
          ],
          "summary": "getHvacSetup",
          "description": "Return Hvac Setup configuration",
          "operationId": "GetHvacSetup",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Hvac Mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HvacSetup"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        },
        "delete": {
          "tags": [
            "setup"
          ],
          "summary": "removeHvacSetup",
          "description": "Remove Hvac Setup configuration. It means the driver Hvac will be blacklisted",
          "operationId": "RemoveHvacSetup",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Hvac Mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
      "/setup/hvac": {
        "post": {
          "tags": [
            "setup"
          ],
          "summary": "setHvacSetup",
          "description": "Sensor Hvac configuration",
          "operationId": "SetHvacSetup",
          "parameters": [],
          "requestBody": {
            "description": "Hvac config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HvacSetup"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HvacSetup"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/config/hvac": {
        "post": {
          "tags": [
            "config"
          ],
          "summary": "setHvacConfig",
          "description": "Change Hvac runtime configuration",
          "operationId": "SetHvacConfig",
          "parameters": [],
          "requestBody": {
            "description": "Hvac runtime config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HvacConfig"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation, Configuration read from database",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/command/hvac": {
        "post": {
          "tags": [
            "command"
          ],
          "summary": "sendHvacCommand",
          "description": "Send Hvac command",
          "operationId": "SendHvacCommand",
          "parameters": [],
          "requestBody": {
            "description": "Hvac Command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HvacCommand"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/status/hvac/{mac}": {
        "get": {
          "tags": [
            "status"
          ],
          "summary": "getHvacStatus",
          "description": "Return the given Hvac status",
          "operationId": "GetHvacStatus",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Hvac Mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HvacStatus"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/commissioning/install": {
        "post": {
          "tags": [
            "commissioning"
          ],
          "summary": "installDriver",
          "description": "Associate a driver to a cable",
          "operationId": "InstallDriver",
          "parameters": [],
          "requestBody": {
            "description": "driver detail description",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverInfo"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BimInfo"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HvacSetup": {
          "title": "HvacSetup",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Hvac mac address"
            },
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "dumpFrequency": {
              "type": "integer",
              "description": "Dump frequency by default: 1s",
              "format": "int32"
            }
          }
        },
        "HvacConfig": {
          "title": "HvacConfig",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Hvac mac address"
            },
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "dumpFrequency": {
              "type": "integer",
              "description": "Dump frequency by default: 1s",
              "format": "int32"
            },
            "isConfigured": {
              "type": "boolean",
              "description": "Is the Hvac configured"
            }
          }
        },
        "HvacCommand": {
          "title": "HvacCommand",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Hvac mac address"
            }
          }
        },
...
        "DriverInfo": {
          "title": "DriverInfo",
          "required": [
            "label",
            "modelName",
            "fullMac"
          ],
          "type": "object",
          "properties": {
            "label": {
              "type": "string",
              "description": "label name available in the IFC for identifying the element"
            },
            "modelName": {
              "type": "string",
              "description": "Model name"
            },
            "fullMac": {
              "type": "string",
              "description": "device mac address (long version)"
            }
          }
        },
...
        "BimInfo": {
          "title": "BimInfo",
          "required": [
            "label"
          ],
          "type": "object",
          "properties": {
            "label": {
              "type": "string",
              "description": "label name available in the IFC for identifying the element"
            },
            "modelName": {
              "type": "string",
              "description": "Model name"
            },
            "mac": {
              "type": "string",
              "description": "device mac address (short version)"
            },
            "fullMac": {
              "type": "string",
              "description": "device mac address (long version)"
            },
            "modbusID": {
              "type": "integer",
              "description": "Modbus unique ID",
              "format": "int32"
            }
          }
        },
...
        "HvacStatus": {
          "title": "HvacStatus",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Hvac mac address"
            },
            "ip": {
              "type": "string",
              "description": "Hvac IP address"
            },
            "group": {
              "type": "integer",
              "description": "Associate group",
              "format": "int32"
            },
            "protocol": {
              "type": "string",
              "description": "Hvac communication protocol"
            },
            "switchMac": {
              "type": "string",
              "description": "Associate switch mac address"
            },
            "isConfigured": {
              "type": "boolean",
              "description": "Is the hvac configured"
            },
            "softwareVersion": {
              "type": "number",
              "description": "hvac Software version"
            },
            "hardwareVersion": {
              "type": "string",
              "description": "hvac Hardware version"
            },
            "error": {
              "type": "integer",
              "description": "Error code value: 0 means no error",
              "format": "int32"
            },
            "friendlyName": {
              "type": "string",
              "description": "Hvac FriendlyName"
            },
            "setpoint": {
              "type": "string",
              "description": "current setpoint"
            },
            "heat": {
              "type": "string",
              "description": "setpoint when the mode is heating"
            },
            "cool": {
              "type": "string",
              "description": "setpoint when the mode is cooling"
            },
            "damper": {
              "type": "string",
              "description": "damper"
            }
          }
        },
...
        "Error": {
          "title": "Error",
          "required": [
            "code",
            "message"
          ],
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "message": {
              "type": "string"
            }
          }
        },
...
      },
      "securitySchemes": {
        "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT",
            "description": "For accessing the API a valid JWT token must be passed in all the queries in the Authorization header. Authorization header : Bearer xxxxxx.yyyyyyy.zzzzzz"
        }
      }
    },
    "security": [
      {
        "Authorization": []
      }
    ],
...
  }